mastersdiving_api/pyproject.toml
[tool.poetry]
name = "mastersdiving-api"
version = "0.1.0"
description = ""
authors = ["Derek Moore <derek@findthebrink.com>"]
license = "UNLICENSED"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
openai = "^1.35.14"
django-filter = "^24.2"
langchain = "^0.2.9"
langchain-openai = "^0.1.17"
langchain-community = "^0.2.7"


[tool.poetry.group.development.dependencies]
ipdb = "^0.13.13"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
